requirements.txt
absl-py>=1.0.0
anyio>=3.4.0
argcomplete>=1.12.3
asgiref>=3.4.1
backcall>=0.2.0
cachetools>=5.0.0
certifi>=2021.10.8
charset-normalizer>=2.0.11
click>=8.0.3
colorama>=0.4.4
cycler>=0.11.0
debugpy>=1.5.1
decorator>=4.4.2
entrypoints>=0.3
fastapi>=0.70.1
fonttools>=4.29.1
google-auth>=2.6.0
google-auth-oauthlib>=0.4.6
grpcio>=1.43.0
gunicorn>=20.1.0
h11>=0.12.0
idna>=3.3
imageio>=2.15.0
imageio-ffmpeg>=0.4.5
importlib-metadata>=4.10.0
ipykernel>=6.6.0
ipython>=7.30.1
jedi>=0.18.1
jupyter-client>=7.1.0
jupyter-core>=4.9.1
jinja2
kiwisolver>=1.3.2
Markdown>=3.3.6
matplotlib>=3.5.1
matplotlib-inline>=0.1.3
moviepy>=1.0.3
nest-asyncio>=1.5.4
numpy>=1.21.5
oauthlib>=3.2.0
opencv-python>=4.5.5.62
opencv-python-headless>=4.5.4.60
packaging>=21.3
pandas>=1.4.0
parso>=0.8.3
pickleshare>=0.7.5
Pillow>=8.4.0
proglog>=0.1.9
prompt-toolkit>=3.0.24
protobuf>=3.19.4
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pydantic>=1.8.2
Pygments>=2.10.0
pyparsing>=3.0.7
python-dateutil>=2.8.2
python-multipart>=0.0.5
pytz>=2021.3
pywin32>=303
PyYAML>=6.0
pyzmq>=22.3.0
requests>=2.27.1
requests-oauthlib>=1.3.1
rsa>=4.8
scipy>=1.8.0
seaborn>=0.11.2
six>=1.16.0
sniffio>=1.2.0
starlette>=0.16.0
tensorboard>=2.8.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.1
torch>=1.10.2
torchvision>=0.11.3
tornado>=6.1
tqdm>=4.62.3
traitlets>=5.1.1
typing-extensions>=4.0.1
urllib3>=1.26.8
uvicorn>=0.16.0
wcwidth>=0.2.5
Werkzeug>=2.0.2
zipp>=3.6.0
